sagemaker/distributed_tabnet/src/requirements_pt.txt
torchmetrics>=0.3 
lightning-bolts==0.5.0
pytorch-lightning<1.7
ray[default]==2.6.3
ray_lightning==0.3.0
pytorch-tabnet==4.0